+++ src/app/destinyTrackerApi/d2-itemTransformer.ts
@@ -6,45 +6,40 @@ import { getPowerMods } from '../inventory/store/d2-item-factory.service';
 import { DtrD2BasicItem, D2ItemFetchRequest } from '../item-review/d2-dtr-api-types';
 
 /**
- * Translates items from the objects that DIM has to the form that the DTR API expects.
- * This is generally tailored towards working with weapons.
- * Also tailored for the Destiny 2 version.
+ * Translate a DIM item into the basic form that the DTR understands an item to contain.
+ * This does not contain personally-identifying information.
+ * Meant for fetch calls.
  */
-class D2ItemTransformer {
-  /**
-   * Translate a DIM item into the basic form that the DTR understands an item to contain.
-   * This does not contain personally-identifying information.
-   * Meant for fetch calls.
-   */
-  translateToDtrItem(item: D2Item | DestinyVendorSaleItemComponent): D2ItemFetchRequest {
-    return {
-      referenceId: ((item as DestinyVendorSaleItemComponent).itemHash !== undefined) ? (item as DestinyVendorSaleItemComponent).itemHash : (item as D2Item).hash
-    };
-  }
-
-  /**
-   * Get the roll and perks for the selected DIM item (to send to the DTR API).
-   * Will contain personally-identifying information.
-   */
-  getRollAndPerks(item: D2Item): DtrD2BasicItem {
-    const powerModHashes = getPowerMods(item).map((m) => m.hash);
-    return {
-      selectedPerks: this._getSelectedPlugs(item, powerModHashes),
-      attachedMods: powerModHashes,
-      referenceId: item.hash,
-      instanceId: item.id,
-    };
-  }
-
-  _getSelectedPlugs(item: D2Item, powerModHashes: number[]) {
-    if (!item.sockets) {
-      return [];
-    }
-
-    const allPlugs = compact(item.sockets.sockets.map((i) => i.plug).map((i) => i && i.plugItem.hash));
-
-    return _.difference(allPlugs, powerModHashes);
-  }
+export function translateToDtrItem(item: D2Item | DestinyVendorSaleItemComponent): D2ItemFetchRequest {
+  return {
+    referenceId: isVendorSaleItem(item) ? item.itemHash : item.hash
+  };
 }
 
-export { D2ItemTransformer };
+/**
+ * Get the roll and perks for the selected DIM item (to send to the DTR API).
+ * Will contain personally-identifying information.
+ */
+export function getRollAndPerks(item: D2Item): DtrD2BasicItem {
+  const powerModHashes = getPowerMods(item).map((m) => m.hash);
+  return {
+    selectedPerks: getSelectedPlugs(item, powerModHashes),
+    attachedMods: powerModHashes,
+    referenceId: item.hash,
+    instanceId: item.id,
+  };
+}
+
+function getSelectedPlugs(item: D2Item, powerModHashes: number[]): number[] {
+  if (!item.sockets) {
+    return [];
+  }
+
+  const allPlugs = compact(item.sockets.sockets.map((i) => i.plug).map((i) => i && i.plugItem.hash));
+
+  return _.difference(allPlugs, powerModHashes);
+}
+
+function isVendorSaleItem(item: D2Item | DestinyVendorSaleItemComponent): item is DestinyVendorSaleItemComponent {
+  return (item as DestinyVendorSaleItemComponent).itemHash !== undefined;
+}