+++ docs/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Next
 
+# 4.53.0 (2018-05-20)
+
 * Add previews for engrams and other preview-able items.
 * Display Crucible ranks on the progress page.
 * Add emotes back to the collections page.